Initial commit
I have been working on this for a while without Git, and it's become large enough to the point where I need to start tracking it.

Forgive the massive flores200_dataset folder, I'll create a setup process which downloads this data rather than storing it directly in the repo. More documentation will come soon.